pyproject.toml
[project]
name = "causal-portfolio-selector"
version = "0.1.0"
description = "Standalone SATzilla-style portfolio selector for causal discovery algorithms"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "joblib>=1.4",
    "networkx>=3.2",
    "numpy>=1.26",
    "pandas>=2.2",
    "pyyaml>=6.0",
    "scikit-learn>=1.5",
    "scipy>=1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
learned = [
    "torch>=2.2",
]
benchmark = [
    "causal-learn>=0.1.3.7",
    "pgmpy>=0.1.25",
]

[project.scripts]
causal-portfolio = "causal_portfolio_selector.cli:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
